Require Forbidden Workshop Level 4 instead of 3 for Short Victory
Fixes the bug for when the player skips level 3, going straight to level 4 from level 2 by requiring level 4 for a short campaign.

In  vco2_main_listeners.lua, there's also this line of code:
`if current_workshop_lvl == 3`

This bug could probably be fixed by making that a `>=` but I didn't want to get all the way to level 3 again just to make sure nothing broke. I know that checking for level 4 works, so I just went with that. Also, it's really easy to get to level 4 (Ikit just has to be level 25, which over the course of a short campaign should easily happen).